+++ module/Application/src/Application/Controller/StatistiqueController.php
@@ -25,8 +25,8 @@ class StatistiqueController extends AbstractController
         $theses = null;
 
         /** TODO prendre les non substitué **/
-        $ecoles = $this->getStructureService()->findAllStructuresAffichablesByType(TypeStructure::CODE_ECOLE_DOCTORALE, 'libelle');
-        $unites = $this->getStructureService()->findAllStructuresAffichablesByType(TypeStructure::CODE_UNITE_RECHERCHE, 'libelle');
+        $ecoles = $this->getStructureService()->findAllStructuresAffichablesByType(TypeStructure::CODE_ECOLE_DOCTORALE, 'structure.libelle');
+        $unites = $this->getStructureService()->findAllStructuresAffichablesByType(TypeStructure::CODE_UNITE_RECHERCHE, 'structure.libelle');
         $etablissements = $this->getEtablissementService()->getRepository()->findAllEtablissementsInscriptions();
 
         /**