+++ module/Application/src/Application/Service/Rapport/RapportSearchService.php
@@ -148,12 +148,12 @@ class RapportSearchService extends SearchService
     private function fetchEcolesDoctorales(): array
     {
         return $this->structureService->findAllStructuresAffichablesByType(
-            TypeStructure::CODE_ECOLE_DOCTORALE, 'sigle', true);
+            TypeStructure::CODE_ECOLE_DOCTORALE, 'structure.sigle');
     }
 
     private function fetchUnitesRecherches(): array
     {
-        return $this->structureService->findAllStructuresAffichablesByType(TypeStructure::CODE_UNITE_RECHERCHE, 'code', false);
+        return $this->structureService->findAllStructuresAffichablesByType(TypeStructure::CODE_UNITE_RECHERCHE, 'structure.code', false);
     }
 
     private function fetchOriginesFinancements(): array