+++ module/Soutenance/src/Soutenance/Service/Proposition/PropositionSearchService.php
@@ -127,13 +127,13 @@ class PropositionSearchService extends SearchService
     private function fetchEcolesDoctorales(SelectSearchFilter $filter): array
     {
         return $this->structureService->findAllStructuresAffichablesByType(
-            TypeStructure::CODE_ECOLE_DOCTORALE, 'sigle', true);
+            TypeStructure::CODE_ECOLE_DOCTORALE, 'structure.sigle');
     }
 
     private function fetchUnitesRecherches(SelectSearchFilter $filter): array
     {
         return $this->structureService->findAllStructuresAffichablesByType(
-            TypeStructure::CODE_UNITE_RECHERCHE, 'code', true);
+            TypeStructure::CODE_UNITE_RECHERCHE, 'structure.code');
     }
 
     /////////////////////////////////////// Sorters /////////////////////////////////////////