+++ module/Structure/src/Structure/Entity/Db/EcoleDoctorale.php
@@ -171,15 +171,13 @@ class EcoleDoctorale implements
     {
         $estFermee = $this->structure->estFermee();
 
-        $subtext = $this->structure->getLibelle();
-        if ($estFermee) {
-            $subtext .= " - FERMÉE";
-        }
+        $libelle = $this->structure->getLibelle();
+        $sigle = $this->structure->getCode() ?: '???';
 
         return [
             'value' => $this->getSourceCode(),
-            'label' => $this->structure->getSigle(),
-            'subtext' => $subtext,
+            'label' => sprintf('%s - %s', $sigle, $libelle),
+            'subtext' => $estFermee ? "Fermée" : null,
             'class' => $estFermee ? 'fermee' : '',
         ];
     }