+++ module/These/src/These/Service/These/TheseSearchService.php
@@ -396,13 +396,13 @@ EOS;
     private function fetchEcolesDoctorales(SelectSearchFilter $filter): array
     {
         return $this->structureService->findAllStructuresAffichablesByType(
-            TypeStructure::CODE_ECOLE_DOCTORALE, 'sigle', true);
+            TypeStructure::CODE_ECOLE_DOCTORALE, 'structure.code');
     }
 
     private function fetchUnitesRecherches(SelectSearchFilter $filter): array
     {
         return $this->structureService->findAllStructuresAffichablesByType(
-            TypeStructure::CODE_UNITE_RECHERCHE, 'code', true);
+            TypeStructure::CODE_UNITE_RECHERCHE, ['structure.sigle', 'structure.libelle']);
     }
 
     private function fetchOriginesFinancements(SelectSearchFilter $filter): array